Serealizaron cambiosen el modulo de Insumos, tales como la presentación, se agrego un campo para contabilizar el inventario del producto

+++ apps/partida/views.py
@@ -67,7 +67,6 @@ def editar(request, idpart):
 def eliminar(request, idpart):
 
 	if request.method == 'POST':
-
 		url_ant = "inicio_partida"
 		reg = PartidasM.objects.get(id=idpart)
 		#se deben eliminar los registros de materiales pertenecientes a esta partida
@@ -93,29 +92,30 @@ def detallar(request, idpart):
 	
 	# ~ print("PARTIDA: ", codp )
 	if request.method =='POST':
-		
 		for i in json.loads(request.POST["ObjDatos"]):
 			
-			if i["accion"] == "eliminar":
-				PartidaDetallesM.objects.filter(idpart=idpart).filter(idism=i["id"]).delete()
-			else:
-				res = PartidaDetallesM.objects.filter(idpart=idpart).filter(idism=i["id"]).exists()
-				if res==True :
-					# ~ El registro existe y se actualiza
-					instance = PartidaDetallesM.objects.filter(idism=i["id"]).update(cant=i["datos"])
+			if i['destino'] == "PARTDETLLS":
+				if i["accion"] == "eliminar":
+					PartidaDetallesM.objects.filter(idpart=idpart).filter(idism=i["id"]).delete()
 				else:
-					# ~ El registro no existe, se crea un nuevo registro
-					g = PartidaDetallesM(idism=i["id"], idpart=idpart, cant=i["datos"])
-					g.save()
-		
+					res = PartidaDetallesM.objects.filter(idpart=idpart).filter(idism=i["id"]).exists()
+					if res==True :
+						# ~ El registro existe y se actualiza
+						instance = PartidaDetallesM.objects.filter(idism=i["id"]).update(cant=i["datos"])
+					else:
+						# ~ El registro no existe, se crea un nuevo registro
+						g = PartidaDetallesM(idism=i["id"], idpart=idpart, cant=i["datos"])
+						g.save()
+			
 		total = 0.0
-		for k in PartidasM.objects.all():
+		# ~ for k in PartidasM.objects.all():
 			# ~ print(k.id)
-			for i in PartidaDetallesM.objects.filter(idpart=k.id):
+		for i in PartidaDetallesM.objects.filter(idpart=idpart):
+			if InsumosM.objects.filter(id=i.idism).exists() == True:
 				ism   = InsumosM.objects.get(id=i.idism)
 				total = total +  ism.cumedida()*i.cant
-			ss = PartidasM.objects.filter(id=k.id).update(cost=round(total,2))
-			total=0.0
+		ss = PartidasM.objects.filter(id=idpart).update(cost=round(total,2))
+		total=0.0
 		
 		url = "/partidas/detallar/" + str(idpart) + "/"
 		# ~ print(url)
@@ -140,7 +140,6 @@ def detallar(request, idpart):
 					 })
 			total = total +  ism.cumedida()*i.cant
 			
-	
 	# ~ agregar el formset a la variable contexto
 	# ~ contexto['formset'] = formset)
 	# ~ print(datos)
